Abstracted opengl init & added types
