+++ fractalizer/src/lib/interfaces.ts
@@ -1,0 +1,7 @@
+export interface ProgramInfo {
+    gl: WebGLRenderingContext | null;
+    zoomLoc: WebGLUniformLocation | null;
+    resolutionLoc: WebGLUniformLocation | null;
+    locationLoc: WebGLUniformLocation | null;
+    fractalTypeLoc: WebGLUniformLocation | null;
+  }
